Discontinuities in a Distribution

Discontinuity points are the coordinates marking the positions of step discontinuities in a distribution. They are the upper points at each discontinuity in the cdf, together with the size of the discontinuity gap.

discontinuities(object)

aggregate_weights(y, weights, sum_to_one = FALSE)

Arguments

object
A distribution object.

y
Vector of outcomes.

weights
Vector of weights, one for each of y. These need not sum to one, but must not be negative.

sum_to_one
Logical; should the weights be normalized to sum to 1? Default is FALSE.

Value

Data frame with the following columns:

location: Increasing vector of unique values of y having positive weight.

size: Weights corresponding to each outcome.

Details

For a vector of outcomes y with a matching vector of weights, aggregate_weights() provides a single non-zero, non-NA weight per unique value of y.

Examples

a <- stepdst(1:5)
discontinuities(a)

#> # A tibble: 5 x 2
#>   location  size
#>      <int> <dbl>
#> 1        1   0.2
#> 2        2   0.2
#> 3        3   0.2
#> 4        4   0.2
#> 5        5   0.2

b <- dst_norm(0, 1)
discontinuities(b)

#> # A tibble: 0 x 2
#> # … with 2 variables: location <dbl>, size <dbl>

c <- graft_right(a, b, sep_y = 3.5)
discontinuities(c)

#> # A tibble: 3 x 2
#>   location  size
#>      <int> <dbl>
#> 1        1   0.2
#> 2        2   0.2
#> 3        3   0.2