distplyr: Manipulate and Combine Probability Distributions

Go beyond standard probability distributions such as the Normal or Exponential by combining, shifting, maximizing, and otherwise transforming distributions with simple, verb-based functions. Provides easy access to a broader space of distributions more representative of real-world systems such as river flows or insurance claims. Part of the probaverse framework of packages to support advanced statistical modeling and simulations with an intuitive workflow.

Details

The distplyr package provides tools for manipulating probability distributions using intuitive syntax. Key features include:

Arithmetic operators: Use +, -, *, /, and ^ with distributions. See Ops.dst() for details.

Mathematical functions: Use log(), exp(), log10(), and sqrt() with distributions. See Math.dst() for details.

Transformation functions: Use shift(), multiply(), flip(), and invert() to transform distributions. See ?linear_transform for details.

Mixing distributions: Use mix() to create mixture distributions.

See also

Author

Maintainer: Vincenzo Coia vincenzo.coia@gmail.com [copyright holder]

Other contributors:

Amogh Joshi [contributor]

Shuyi Tan [contributor]

Zhipeng Zhu [contributor]